+++ components/vf-banner/CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2.0.4
 
 * Dependency update to be compatible with ESM packages[Tracking issue](https://github.com/visual-framework/vf-core/issues/2107)
+* Changed: Test cases added for Angular  [Tracking issue](https://github.com/visual-framework/vf-core/issues/1932)
 
 ### 2.0.3